requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.1.4
django-ratelimit==4.1.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kombu==5.4.2
openai==1.57.4
prompt_toolkit==3.0.48
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
